Cheaper headers
ClangBuildAnalyzer
  Parsing (frontend):         1545.4 s
  Codegen & opts (backend):    484.0 s
