+++ src/query/procedure/cypher_types.hpp
@@ -1,4 +1,4 @@
-// Copyright 2022 Memgraph Ltd.
+// Copyright 2024 Memgraph Ltd.
 //
 // Use of this software is governed by the Business Source License
 // included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
@@ -12,17 +12,17 @@
 /// @file
 #pragma once
 
-#include "mg_procedure.h"
-
-#include <functional>
-#include <memory>
-#include <string_view>
-
 #include "query/procedure/cypher_type_ptr.hpp"
 #include "query/procedure/mg_procedure_impl.hpp"
 #include "query/typed_value.hpp"
 #include "utils/memory.hpp"
 #include "utils/pmr/string.hpp"
+
+#include "mg_procedure.h"
+
+#include <functional>
+#include <memory>
+#include <string_view>
 
 namespace memgraph::query::procedure {
