+++ src/storage/v2/disk/edge_type_index.cpp
@@ -13,6 +13,8 @@
 
 #include "utils/exceptions.hpp"
 
+#include "spdlog/spdlog.h"
+
 namespace memgraph::storage {
 
 bool DiskEdgeTypeIndex::DropIndex(EdgeTypeId /*edge_type*/) {
