+++ src/storage/v2/disk/label_property_index.cpp
@@ -12,6 +12,7 @@
 /// TODO: clear dependencies
 
 #include "storage/v2/disk/label_property_index.hpp"
+#include "storage/v2/transaction.hpp"
 #include "utils/disk_utils.hpp"
 #include "utils/rocksdb_serialization.hpp"
