+++ src/storage/v2/edge_ref.cpp
@@ -9,26 +9,14 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-#pragma once
+#include "storage/v2/edge_ref.hpp"
 
-#include "storage/v2/id_types.hpp"
+#include <type_traits>
 
 namespace memgraph::storage {
 
-// Forward declaration because we only store a pointer here.
-struct Edge;
-
-struct EdgeRef {
-  explicit EdgeRef(Gid gid) : gid(gid) {}
-  explicit EdgeRef(Edge *ptr) : ptr(ptr) {}
-
-  friend bool operator==(const EdgeRef &a, const EdgeRef &b) noexcept { return a.gid == b.gid; }
-  friend bool operator<(const EdgeRef &first, const EdgeRef &second) { return first.gid < second.gid; }
-
-  union {
-    Gid gid;
-    Edge *ptr;
-  };
-};
-
+static_assert(sizeof(Gid) == sizeof(Edge *), "The Gid should be the same size as an Edge *!");
+static_assert(std::is_standard_layout_v<Gid>, "The Gid must have a standard layout!");
+static_assert(std::is_standard_layout_v<Edge *>, "The Edge * must have a standard layout!");
+static_assert(std::is_standard_layout_v<EdgeRef>, "The EdgeRef must have a standard layout!");
 }  // namespace memgraph::storage
