+++ src/storage/v2/indices/point_index.hpp
@@ -14,96 +14,9 @@
 #include "storage/v2/indices/point_index_change_collector.hpp"
 #include "utils/skip_list.hpp"
 
-#include <boost/geometry.hpp>
-#include <boost/geometry/geometries/point.hpp>
-#include <boost/geometry/index/rtree.hpp>
-
-namespace bg = boost::geometry;
-namespace bgi = boost::geometry::index;
-
 namespace memgraph::storage {
 
-struct IndexPointWGS2d {
-  explicit IndexPointWGS2d(Point2d point) : rep{point.x(), point.y()} { DMG_ASSERT(IsWGS(point.crs())); }
-  using point_type = bg::model::point<double, 2, bg::cs::spherical_equatorial<bg::degree>>;
-  point_type rep;
-};
-struct IndexPointWGS3d {
-  explicit IndexPointWGS3d(Point3d point) : rep{point.x(), point.y(), point.z()} { DMG_ASSERT(IsWGS(point.crs())); }
-  using point_type = bg::model::point<double, 3, bg::cs::spherical_equatorial<bg::degree>>;
-  point_type rep;
-};
-struct IndexPointCartesian2d {
-  explicit IndexPointCartesian2d(Point2d point) : rep{point.x(), point.y()} { DMG_ASSERT(IsCartesian(point.crs())); }
-  using point_type = bg::model::point<double, 2, bg::cs::cartesian>;
-  point_type rep;
-};
-struct IndexPointCartesian3d {
-  explicit IndexPointCartesian3d(Point3d point) : rep{point.x(), point.y(), point.z()} {
-    DMG_ASSERT(IsCartesian(point.crs()));
-  }
-  using point_type = bg::model::point<double, 3, bg::cs::cartesian>;
-  point_type rep;
-};
-
-template <typename Point>
-struct Entry {
-  using point_type = typename Point::point_type;
-  Entry(Point p, Vertex const *vertex) : p_(p), vertex_(vertex) {}
-
-  friend bool operator==(Entry const &lhs, Entry const &rhs) {
-    if (lhs.vertex_ != rhs.vertex_) return false;
-    if (!boost::geometry::equals(lhs.p_, rhs.p_)) return false;
-    return true;
-  };
-
-  auto point() const -> point_type const & { return p_.rep; }
-  auto vertex() const -> storage::Vertex const * { return vertex_; }
-
- private:
-  Point p_;
-  storage::Vertex const *vertex_;
-};
-};  // namespace memgraph::storage
-
-template <typename IndexPoint>
-struct bg::index::indexable<memgraph::storage::Entry<IndexPoint>> {
-  using result_type = typename IndexPoint::point_type;
-  auto operator()(memgraph::storage::Entry<IndexPoint> const &val) const -> result_type const & { return val.point(); }
-};
-
-namespace memgraph::storage {
-
-template <typename IndexPoint>
-using index_t = bgi::rtree<Entry<IndexPoint>, bgi::quadratic<64>>;  // TODO: tune this
-
-// TODO: use PIMPL idiom to hide the detail that we are using boost/geometry*
-struct PointIndex {
-  PointIndex() = default;
-
-  PointIndex(std::span<Entry<IndexPointWGS2d>> points2dWGS, std::span<Entry<IndexPointCartesian2d>> points2dCartesian,
-             std::span<Entry<IndexPointWGS3d>> points3dWGS, std::span<Entry<IndexPointCartesian3d>> points3dCartesian);
-
-  auto CreateNewPointIndex(LabelPropKey labelPropKey, absl::flat_hash_set<Vertex const *> const &changed_vertices) const
-      -> PointIndex;
-
-  auto EntryCount() const -> std::size_t {
-    return wgs_2d_index_->size() + wgs_3d_index_->size() + cartesian_2d_index_->size() + cartesian_3d_index_->size();
-  }
-
- private:
-  PointIndex(std::shared_ptr<index_t<IndexPointWGS2d>> points2dWGS,
-             std::shared_ptr<index_t<IndexPointWGS3d>> points3dWGS,
-             std::shared_ptr<index_t<IndexPointCartesian2d>> points2dCartesian,
-             std::shared_ptr<index_t<IndexPointCartesian3d>> points3dCartesian);
-
-  std::shared_ptr<index_t<IndexPointWGS2d>> wgs_2d_index_ = std::make_shared<index_t<IndexPointWGS2d>>();
-  std::shared_ptr<index_t<IndexPointWGS3d>> wgs_3d_index_ = std::make_shared<index_t<IndexPointWGS3d>>();
-  std::shared_ptr<index_t<IndexPointCartesian2d>> cartesian_2d_index_ =
-      std::make_shared<index_t<IndexPointCartesian2d>>();
-  std::shared_ptr<index_t<IndexPointCartesian3d>> cartesian_3d_index_ =
-      std::make_shared<index_t<IndexPointCartesian3d>>();
-};
+struct PointIndex;
 
 using index_container_t = std::map<LabelPropKey, std::shared_ptr<PointIndex const>>;
