+++ tests/unit/typed_value.cpp
@@ -13,9 +13,6 @@
 // Copyright 2017 Memgraph
 // Created by Florijan Stamenkovic on 24.01.17..
 //
-#include <functional>
-#include <map>
-#include <set>
 #include <vector>
 
 #include "gtest/gtest.h"
